+++ src/main/java/cn/taketoday/context/utils/DataSize.java
@@ -109,6 +109,20 @@ public final class DataSize implements Comparable<DataSize> {
      *            the amount of the size, measured in terms of the unit, positive or
      *            negative
      * @return a corresponding {@link DataSize}
+     * @since 2.1.6
+     */
+    public static DataSize of(long amount) {
+        return of(amount, DataUnit.BYTES);
+    }
+
+    /**
+     * Obtain a {@link DataSize} representing an amount in the specified
+     * {@link DataUnit}.
+     * 
+     * @param amount
+     *            the amount of the size, measured in terms of the unit, positive or
+     *            negative
+     * @return a corresponding {@link DataSize}
      */
     public static DataSize of(long amount, DataUnit unit) {
         return new DataSize(Math.multiplyExact(amount, unit.size().toBytes()));
@@ -241,15 +255,7 @@ public final class DataSize implements Comparable<DataSize> {
 
     @Override
     public boolean equals(Object other) {
-
-        if (this == other) {
-            return true;
-        }
-
-        if (other instanceof DataSize) {
-            return this.bytes == ((DataSize) other).bytes;
-        }
-        return false;
+        return this == other || (other instanceof DataSize && this.bytes == ((DataSize) other).bytes);
     }
 
     @Override